requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   starlette
    #   watchfiles
babel==2.17.0
    # via mkdocs-material
backrefs==5.9
    # via mkdocs-material
bandit==1.8.6
    # via -r requirements.in
black==25.1.0
    # via -r requirements.in
certifi==2025.7.14
    # via requests
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   black
    #   mkdocs
    #   uvicorn
colorama==0.4.6
    # via mkdocs-material
coverage[toml]==7.10.1
    # via pytest-cov
distlib==0.4.0
    # via virtualenv
fastapi==0.116.1
    # via -r requirements.in
filelock==3.18.0
    # via virtualenv
flake8==7.3.0
    # via -r requirements.in
ghp-import==2.1.0
    # via mkdocs
h11==0.16.0
    # via uvicorn
httptools==0.6.4
    # via uvicorn
identify==2.6.12
    # via pre-commit
idna==3.10
    # via
    #   anyio
    #   requests
iniconfig==2.1.0
    # via pytest
isort==6.0.1
    # via -r requirements.in
jinja2==3.1.6
    # via
    #   -r requirements.in
    #   mkdocs
    #   mkdocs-material
markdown==3.8.2
    # via
    #   mkdocs
    #   mkdocs-material
    #   pymdown-extensions
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   jinja2
    #   mkdocs
mccabe==0.7.0
    # via flake8
mdurl==0.1.2
    # via markdown-it-py
mergedeep==1.3.4
    # via
    #   mkdocs
    #   mkdocs-get-deps
mkdocs==1.6.1
    # via
    #   -r requirements.in
    #   mkdocs-material
mkdocs-get-deps==0.2.0
    # via mkdocs
mkdocs-material==9.6.16
    # via -r requirements.in
mkdocs-material-extensions==1.3.1
    # via mkdocs-material
mypy==1.17.1
    # via -r requirements.in
mypy-extensions==1.1.0
    # via
    #   black
    #   mypy
nodeenv==1.9.1
    # via pre-commit
packaging==25.0
    # via
    #   black
    #   mkdocs
    #   pytest
paginate==0.5.7
    # via mkdocs-material
pathspec==0.12.1
    # via
    #   black
    #   mkdocs
    #   mypy
pbr==6.1.1
    # via stevedore
platformdirs==4.3.8
    # via
    #   black
    #   mkdocs-get-deps
    #   virtualenv
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
pre-commit==4.2.0
    # via -r requirements.in
pycodestyle==2.14.0
    # via flake8
pydantic==2.11.7
    # via
    #   -r requirements.in
    #   fastapi
pydantic-core==2.33.2
    # via pydantic
pyflakes==3.4.0
    # via flake8
pygments==2.19.2
    # via
    #   mkdocs-material
    #   pytest
    #   rich
pymdown-extensions==10.16.1
    # via mkdocs-material
pytest==8.4.1
    # via
    #   -r requirements.in
    #   pytest-cov
pytest-cov==6.2.1
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via ghp-import
python-dotenv==1.1.1
    # via
    #   -r requirements.in
    #   uvicorn
pyyaml==6.0.2
    # via
    #   -r requirements.in
    #   bandit
    #   mkdocs
    #   mkdocs-get-deps
    #   pre-commit
    #   pymdown-extensions
    #   pyyaml-env-tag
    #   uvicorn
pyyaml-env-tag==1.1
    # via mkdocs
requests==2.32.4
    # via mkdocs-material
rich==14.1.0
    # via bandit
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
starlette==0.47.2
    # via fastapi
stevedore==5.4.1
    # via bandit
typing-extensions==4.14.1
    # via
    #   anyio
    #   fastapi
    #   mypy
    #   pydantic
    #   pydantic-core
    #   starlette
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
urllib3==2.5.0
    # via requests
uvicorn[standard]==0.35.0
    # via -r requirements.in
uvloop==0.21.0
    # via uvicorn
virtualenv==20.32.0
    # via pre-commit
watchdog==6.0.0
    # via mkdocs
watchfiles==1.1.0
    # via uvicorn
websockets==15.0.1
    # via uvicorn

# The following packages are considered to be unsafe in a requirements file:
# setuptools
